damage template from roger
in millimeter

from   Roger McNichols rmcnichols@visualaseinc.com
with the larger fiber (say, in liver or prostate) we can create up to 3.5 x 3 cm lesions.
That should be enough. With the smaller fiber probably 2.5 x 2 cm is close to max size,
but with reeeaaaallly long exposures, it might grow larger
25mm x20mm
 scale transverse radius to  10.0mm = 2.0cm diameter